feat(tests): 20 unit tests + fix all clippy warnings (-D warnings clean)
Tests added (inline #[cfg(test)] modules):
- settings: round-trip, escape/unescape, strip_quotes, comment/blank
  handling, hotkey range guard, unicode preservation (8 tests)
- main: compose_final_text pure helper — prefix prepend, no
  double-prefix, truncated flag precedence, unicode bodies (6 tests)
- tray/linux: format_log_label pure helper — Today/Yesterday/ISO date
  formatting, malformed filename fallback, short-date bounds (6 tests)

Warnings: 27 → 0

- Allow non_upper_case_globals at src/hotkey/linux.rs module level
  (X11 XK_End / KeyPress constants are upstream CamelCase)
- Allow dead_code crate-wide — legitimately retained API surface
  (Settings::config_dir, logging::is_enabled, transcribe::load_wav)
  for tests and cross-platform ports
- drop(engine.take()) in worker — explicit free before reload,
  silences assigned_never_read
- Delete unused Tray::set_hotkey_sender (private_in_public)
- Remove unused mpsc import in hotkey/mod.rs
- BufferFullCallback type alias (very_complex_type)
- Drop redundant i32 casts in hotkey/linux.rs grab calls
- Settings::load uses struct-init not Default + field-assign
- Collapse nested if in cleanup_old_wavs

Also extracted compose_final_text and format_log_label as pure
functions so the test surface can grow without needing GTK/enigo.

+++ src/hotkey/linux.rs
@@ -1,3 +1,10 @@
+// X11 keysym and xlib constants ship CamelCase / Mixed_Case (e.g. XK_End,
+// KeyPress, Mod5Mask) because they match the historical Xlib C headers.
+// Clippy flags these as "constant in pattern should have an upper case name"
+// whenever we match against them. They're re-exported verbatim by the x11
+// crate and we can't rename them — allow the lint at module level.
+#![allow(non_upper_case_globals)]
+
 use super::{HotkeyCmd, KeyEvent};
 use std::sync::mpsc;
 use std::thread;
@@ -52,8 +59,8 @@ unsafe fn grab_key(display: *mut Display, root: Window, keycode: u32) {
             m,
             root,
             1,            // owner_events = True
-            GrabModeAsync as i32,
-            GrabModeAsync as i32,
+            GrabModeAsync,
+            GrabModeAsync,
         );
     }
     XSync(display, 0);
